Fix : Change Long variable on Integer / adding FeignClient

+++ Back-end/src/main/java/com/simplon/dvdStore/controllers/DvdController.java
@@ -20,24 +20,29 @@ public class DvdController {
     private final DvdMapper dvdMapper = DvdMapper.INSTANCE;
     @Autowired
     private DvdService service;
+
     @PostMapping
     public void save(@RequestBody DvdDTO dvdDTO) {
         service.save(dvdMapper.dvdDTOToDvdService(dvdDTO));
     }
+
     @GetMapping
     public List<DvdGetDTO> findAll() {
         return dvdMapper.listDvdServiceToDvdGetDTO(service.findAll());
     }
+
     @GetMapping("{id}")
-    public DvdGetDTO findById(@PathVariable long id) {
+    public DvdGetDTO findById(@PathVariable int id) {
         return dvdMapper.dvdServiceToDvdGetDTO(service.findById(id));
     }
+
     @PutMapping("{id}")
-    public void update(@PathVariable long id, @RequestBody DvdDTO dvdDTO) {
+    public void update(@PathVariable int id, @RequestBody DvdDTO dvdDTO) {
         service.update(dvdMapper.dvdDTOToDvdService(dvdDTO), id);
     }
+
     @DeleteMapping("{id}")
-    public void deleteById(@PathVariable long id) {
+    public void deleteById(@PathVariable int id) {
         service.deleteById(id);
     }
 }
